requirements.txt
requests
jinja2
tzlocal
python-dotenv

